services/meds/pyproject.toml
[tool.poetry]
name = "meds"
version = "0.1.0"
description = "Meds OCR microservice"
authors = ["Kilo <kilo@example.com>"]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
fastapi = "^0.100"
uvicorn = "^0.22"
sqlmodel = "^0.0.8"
pytesseract = "^0.3.10"
pillow = "^10.0.0"
opencv-python = "^4.8.0"
httpx = "^0.27.0"
python-multipart = "^0.0.9"

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"
